Revert "motion model"
