+++ Assignment/06/code/gaussian.py
@@ -10,39 +10,9 @@ from matplotlib import pyplot as plt
 from pyexcel_ods import get_data
 from scipy.stats import norm
 from matplotlib.mlab import PCA
-from scipy import optimize
 
-
-def gaussian_func(height, center_x, center_y, width_x, width_y):
-    """Returns a gaussian function with the given parameters"""
-    width_x = float(width_x)
-    width_y = float(width_y)
-    return lambda x,y: height*np.exp(
-                -(((center_x-x)/width_x)**2+((center_y-y)/width_y)**2)/2)
-
-def moments(data):
-    """Returns (height, x, y, width_x, width_y)
-    the gaussian parameters of a 2D distribution by calculating its
-    moments """
-    total = data.sum()
-    X, Y = np.indices(data.shape)
-    x = (X*data).sum()/total
-    y = (Y*data).sum()/total
-    col = data[:, int(y)]
-    width_x = np.sqrt(np.abs((np.arange(col.size)-y)**2*col).sum()/col.sum())
-    row = data[int(x), :]
-    width_y = np.sqrt(np.abs((np.arange(row.size)-x)**2*row).sum()/row.sum())
-    height = data.max()
-    return height, x, y, width_x, width_y
-
-def fitgaussian(data):
-    """Returns (height, x, y, width_x, width_y)
-    the gaussian parameters of a 2D distribution found by a fit"""
-    params = moments(data)
-    errorfunction = lambda p: np.ravel(gaussian_func(*p)(*np.indices(data.shape)) -
-                                 data)
-    p, success = optimize.leastsq(errorfunction, params)
-    return p
+def gaussian(x, mu, sig):
+    return np.exp(-np.power(x - mu, 2.) / (2 * np.power(sig, 2.)))
 
 
                                  
@@ -68,40 +38,40 @@ def plot_hist(data,nbins):
     #plt.show()
     #plt.savefig(path)
 
-def do_pca(data,index):
+def do_pca(data):
     pcaObject = PCA(data)
     #centered_data = pcaObject.center(data)
     projected_data = pcaObject.project(data)
     #print projected_data
-    plot_hist(projected_data[:,index],8)
+    plot_hist(projected_data[:,1],8)
     plt.show()
     
-if __name__ == "__main__":
-    #convert from ordered dict to lists
-    straight_motion_data =  get_data("../measurements2/second_see_data_original.ods",start_row=5,
-                                     row_limit=40,start_column=21,column_limit=4)
-                                     
-    data = straight_motion_data.get("Sheet1")
-    #dataMatrix = np.array(data)
-    
-    data_front_x = list()
-    data_front_y = list()
-    data_back_x = list()
-    data_back_y = list()
-    for x in range(0,40):
-        temp = data[x]
-        data_front_x.append(temp[0])
-        data_front_y.append(temp[1])
-        data_back_x.append(temp[2])
-        data_back_y.append(temp[3])
-    
-    robot_center_x = np.asarray(data_back_x) + (( np.asarray(data_front_x) -  np.asarray(data_back_x))/2)
-    robot_center_y = np.asarray(data_back_y) + (( np.asarray(data_front_y) -  np.asarray(data_back_y))/2)
-    robot_pose_angular = np.round(np.degrees(np.arctan2(robot_center_x,robot_center_y)),2)
-    robot_pose_linear = np.round(np.sqrt(robot_center_x*robot_center_x + robot_center_y*robot_center_y),2)
-    
-    data = np.c_[robot_pose_linear,robot_pose_angular]
-    do_pca(data,0)
+
+#convert from ordered dict to lists
+straight_motion_data =  get_data("../measurements2/second_see_data_original.ods",start_row=5,
+                                 row_limit=40,start_column=21,column_limit=4)
+                                 
+data = straight_motion_data.get("Sheet1")
+#dataMatrix = np.array(data)
+
+data_front_x = list()
+data_front_y = list()
+data_back_x = list()
+data_back_y = list()
+for x in range(0,40):
+    temp = data[x]
+    data_front_x.append(temp[0])
+    data_front_y.append(temp[1])
+    data_back_x.append(temp[2])
+    data_back_y.append(temp[3])
+
+robot_center_x = np.asarray(data_back_x) + (( np.asarray(data_front_x) -  np.asarray(data_back_x))/2)
+robot_center_y = np.asarray(data_back_y) + (( np.asarray(data_front_y) -  np.asarray(data_back_y))/2)
+robot_pose_angular = np.round(np.degrees(np.arctan2(robot_center_x,robot_center_y)),2)
+robot_pose_linear = np.round(np.sqrt(robot_center_x*robot_center_x + robot_center_y*robot_center_y),2)
+
+data = np.c_[robot_pose_linear,robot_pose_angular]
+do_pca(data)
 
 #path = "/home/chaitanya/MAS3/SEE-Project/Assignment/03/diagrams"
 #plot_hist(robot_center_y,4)
