+++ Assignment/07/latex/SEEHomework_Chaitany_Hebbal_Jose_Mayoral_071116.tex
@@ -6,7 +6,6 @@
 \usepackage{nameref}
 \usepackage{hyperref}
 \usepackage{color,soul}
-\usepackage{listings}
 
 \title{Scientific Experimentation and Evaluation\\
 		\small{Assignment 7}}
@@ -15,7 +14,7 @@
 
 \begin{document}
 	\maketitle
-\section{Task Description}
+\section*{Task Description}
 
 Using a modified LEGO robot from the experiment \textit{"Manual motion observation"} a new set of poses will be acquired via AICISS lab optical localization system, then the data will be analyzed to observe the variation on the poses and look at the error distributions in order to observe the variation on the measures and look at the error distributions.
 
@@ -25,7 +24,7 @@ The layout of this experiment consists on:
 	\item The Device Under Test is a modified LEGO robot from the experiment \textit{"Manual motion observation"}.
 \end{enumerate}
 
-\section{AICISS Lab Optical Localization System}
+\section*{AICISS Lab Optical Localization System}
 
 A optical-based localization system will be used in this experiment. From \cite{OptSystem} the optical systems have several advantage to take into account:
 
@@ -69,9 +68,9 @@ Speaking about software, it has some components that need to be run in the works
 %\label{fig:visualizator}
 %\end{figure}
 
-\section{Experiment setup}
+\section*{Experiment setup}
 
-\subsection{Device Under Test}
+\subsection*{Device Under Test}
 
 A tracker have been added to the top of the robot in order to make able the system to track it, the robot configuration is shown at figure \ref{fig:1}.
 
@@ -82,7 +81,7 @@ A tracker have been added to the top of the robot in order to make able the syst
 \label{fig:1}
 \end{figure}
 
-\subsection{Measuring Method}
+\subsection*{Measuring Method}
 
 The measuring method for each curve consists on:
 \begin{enumerate}
@@ -104,7 +103,7 @@ The measuring method for each curve consists on:
 \end{figure}
 
 
-\subsection{What difficulties are expected?}
+\subsection*{What difficulties are expected?}
 
 \begin{itemize}
 	\item Other tracker in the arena must be hidden.
@@ -112,7 +111,7 @@ The measuring method for each curve consists on:
 	\item Tracker can loses the robot for instances then the data is not noise-free.
 \end{itemize}
 
-\subsection{Data Processing}
+\subsection*{Data Processing}
 
 The raw measurements are shown at figure \ref{fig:realScaleraw}. This figure doesn't show any kind of data processing so it makes able to see some noisy measurements that don't relies to any of the five curves. Also the figure shows an unexpected problem with a measurement method which is that the curve were acquired in two different days then the initial marker were lost for the second day. In the figure, straight and slightleft curve were taken in the first day, the rest on the second one.
 
@@ -125,7 +124,7 @@ The raw measurements are shown at figure \ref{fig:realScaleraw}. This figure doe
 
 %However the figure \ref{fig:rawMeasurements} doesn't show the real size of the environment. The correct scale is shown at \ref{fig:realScale}.
 
-\begin{figure}[H]
+\begin{figure}[ht!]
 \centering
 \includegraphics[scale=0.50]{images/realScaleraw}
 \caption{Raw Acquired Data}
@@ -145,7 +144,7 @@ Then this calculated initial pose was substracted to all the poses from the file
 \]
 Additionally the sizes of the measurements files were trim to the minimum length of all of the measurements. The result project all measurements to the origin of a frame \ref{fig:preprocess}.
 
-\begin{figure}[H]
+\begin{figure}[ht!]
 \centering
 \includegraphics[trim={220 0 0 0},clip,scale=0.5]{images/centerMeasurements}
 \caption{CenterData}
@@ -157,23 +156,16 @@ The new set of measurements are as shown in Figure:\ref{fig:OriginalData}
 
 \begin{figure}[ht!]
 \centering
-\includegraphics[scale=0.25]{images/figure_1}
+\includegraphics[scale=0.6]{images/figure_1}
 \caption{New set of measurements}
 \label{fig:OriginalData}
 \end{figure}
 
 \newpage
 
-\subsection{Motion Model}
+\subsection*{\hl{Motion Model}}
 
-\begin{figure}[H]
-\centering
-\includegraphics[scale=0.2]{images/velocity-motion-model}
-\caption{Velocity motion model \cite{MotionModel}}
-\label{fig:motionmodel}
-\end{figure}
-
-We use the velocity motion model as shown in Figure:\ref{fig:motionmodel} According to this model, the probability of a robot’s pose at time $t$ given a motion command and the pose at time $t − 1$ is given as in \cite{Alex}
+We use the velocity motion model specified in the book "Probablistic robotics" by sebastian thrun.According to this model, the probability of a robot’s pose at time $t$ given a motion command and the pose at time $t − 1$ is given as in \cite{Alex}
 
 \[
 	P(x_t|x_{t-1},u_t) = N(0,\alpha_1 v^2+\alpha_2 \omega^2) \cdot
@@ -191,58 +183,22 @@ We made use of the code provided in the sample solution \cite{Alex} to estimate 
 
 The values obtained are as follows: 
 
-$\alpha_1$ =  9.36575693e-06 
+$\alpha_1$ =  3.17679348e-06 
 
-$\alpha_2$ =  8.53840417e-06
+$\alpha_2$ =  2.61282443e-06
 
-$\alpha_3$ =  9.22593223e-06
+$\alpha_3$ =  7.33286509e-06
 
-$\alpha_4$ =  6.02069757e-06
+$\alpha_4$ =  4.43967131e-06
 
-$\alpha_5$ =  2.54877909e-06
+$\alpha_5$ =  6.38137103e-06
 
-$\alpha_6$ =  8.97275720e-06
-     
+$\alpha_6$ =  9.99649623e-06
+        
+      
+\subsection*{Parameters used to drive the robot}
 
-Once we have estimated the values of alphas, we sample data using following algorithm as shown in Figure:\ref{fig:samplingmodel}
-
-\begin{figure}[ht!]
-\centering
-\includegraphics[scale=0.2]{images/sampling-model}
-\caption{Sample motion model velocity \cite{MotionModel}}
-\label{fig:samplingmodel}
-\end{figure}
-
-\subsection{Results}
-
-\begin{figure}[H]
-\centering
-\includegraphics[scale=0.25]{images/figure_1}
-\caption{Original Data}
-\label{fig:OriginalData}
-\end{figure}
-
-\begin{figure}[H]
-\centering
-\includegraphics[scale=0.25]{images/figure_2}
-\caption{Sampled Data}
-\label{fig:SampledData}
-\end{figure}        
-
-From figures\ref{fig:OriginalData} and \ref{fig:SampledData} it is clear that the motion model perfectly fits the observed data for the estimated values of alpha.
-
-\begin{thebibliography}{1}
-\bibitem{OptSystem} Optical Tracking Explanation: \href{http://www.ps-tech.com/3d-technology/optical-tracking}{Link}
-\bibitem{Alex} Motion Model Parameter Estimation:Sample Solution
-\bibitem{MotionModel} Probablistic Motion Models by Wolfram Burgard, Cyrill Stachniss,Maren Bennewitz, Kai Arras:   \href{http://ais.informatik.uni-freiburg.de/teaching/ss11/robotics/slides/06-motion-models.pdf} {Link}
-\end{thebibliography}
-
-
-\section{Appendix}
-
-\subsection{Parameters used to drive the robot}
-
-\begin{table}[h!]
+\begin{table}[ht]
 \centering
 \caption{Parameters Used}
 \label{tab:1}
@@ -257,27 +213,13 @@ Straight        & 160        & 90    &          & 12 \\ \hline
 \end{tabular}
 \end{table}
 
-\newpage
-\subsection{Sampling code}
 
-\begin{lstlisting}
- 
-def sample_motion_model_velocity(ut,x_prevt,alphas,delta_t):
-    var1 = alphas[0]*(ut[0]*ut[0]) + alphas[1]*(ut[1]*ut[1])
-    var2 = alphas[2]*(ut[0]*ut[0]) + alphas[3]*(ut[1]*ut[1])
-    var3 = alphas[4]*(ut[0]*ut[0]) + alphas[5]*(ut[1]*ut[1])
-    v_hat = ut[0] + np.random.normal(0,var1)
-    omega_hat = ut[1] + np.random.normal(0,var2)
-    gamma_hat = 0#np.random.normal(0,var3)
-    K = (v_hat/omega_hat)
-    xt = x_prevt[0] - K * np.sin(x_prevt[2]) + K * np.sin(x_prevt[2] + (omega_hat*delta_t))
-    yt = x_prevt[1] + K * np.cos(x_prevt[2]) - K * np.cos(x_prevt[2] + (omega_hat*delta_t))
-    theta = x_prevt[2] + (omega_hat*delta_t) +(gamma_hat*delta_t)
-    
-    return np.array([xt,yt,theta])
-    
-\end{lstlisting}
+\section*{Conclusions}
 
 
+\begin{thebibliography}{1}
+\bibitem{OptSystem} Optical Tracking Explanation \href{http://www.ps-tech.com/3d-technology/optical-tracking}{Link}
+\bibitem{Alex} Motion Model Parameter Estimation:Sample Solution
+\end{thebibliography}
 
 \end{document}
