--- Assignment/07/latex/missfont.log
@@ -1,4 +1,0 @@
-mktextfm ectt1000
-mktextfm ectt1000
-mktextfm ectt1000
-mktextfm ectt1000
