--- Assignment/observed_motion.py
@@ -1,10 +1,0 @@
-# -*- coding: utf-8 -*-
-"""
-Created on Wed Nov 16 14:22:21 2016
-
-@author: chaitanya
-"""
-
-import numpy as np
-from calculate_alphas import data_read
-
